Change member variables from uint8_t and int to uint

+++ lib/ble_packet_sink_impl.h
@@ -19,15 +19,15 @@ class ble_packet_sink_impl : public ble_packet_sink
 private:
     uint8_t slice(float data_in); // Slice float data into binary data
 
-    uint8_t d_access_code_len = 32; // Length of the code
-    uint64_t d_shift_reg = 0;       // Shift register for the incoming data
-    uint8_t d_fill_buffer = 0; // Ensures the shift register is filled before comparing
-    uint64_t d_access_code;    // Access code to be detected
-    uint64_t d_mask;           // (1 << code_len) - 1
-    uint8_t d_threshold;       // Allowed bit errors in the preamble detection
+    uint d_access_code_len = 32; // Length of the code
+    uint64_t d_shift_reg = 0;    // Shift register for the incoming data
+    uint d_fill_buffer = 0;      // Ensures the shift register is filled before comparing
+    uint64_t d_access_code;      // Access code to be detected
+    uint64_t d_mask;             // (1 << code_len) - 1
+    uint d_threshold;            // Allowed bit errors in the preamble detection
 
 public:
-    ble_packet_sink_impl(uint32_t base_address, int preamble_threshold);
+    ble_packet_sink_impl(uint32_t base_address, uint preamble_threshold);
     ~ble_packet_sink_impl();
 
     // Where all the action really happens
